adds search functionality
