+++ src/SearchContainer.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react'
+import Search from './Search.js'
 import stockData from '../data/stock-data.json'
 import axios from 'axios'
 
@@ -7,7 +8,7 @@ class SearchContainer extends Component {
 		super(props);
 		this.state = {
 			stockDataAPI: stockData,
-			stockSymbol: this.props.match.params.symbol
+			symbol: ''
 		}
 		this.handleSubmit = this.handleSubmit.bind(this)
 		this.searchStock = this.searchStock.bind(this)
@@ -15,13 +16,25 @@ class SearchContainer extends Component {
 	}
 	searchStock() {
 		console.log("search")
+		axios.get(`http://finance.google.com/finance/info?client=ig&q=NASDAQ%3A${this.state.symbol}`).then((response) => {
+			let blah = JSON.parse(response.data.slice(3));
+			console.log(blah)
+			this.setState({
+				stockDataAPI: JSON.parse(response.data.slice(3))
+			})
+		}).catch((err) => {
+			console.log(err)
+		})
 	}
 	handleSubmit(event) {
 		this.searchStock()
 		event.preventDefault()
 	}
 	handleChange(event) {
-		console.log("change")
+		const name = event.target.name
+		this.setState({
+			[name]: event.target.value
+		}, () => { console.log(this.state)})
 	}
 	render() {
 		return (
@@ -29,6 +42,7 @@ class SearchContainer extends Component {
 				<h1>Search Symbols</h1>
 	        	<input type="text" name="symbol" placeholder="Stock Symbol" onChange={this.handleChange}/>
 	        	<input type="submit" value="Submit" onClick={this.handleSubmit}/>
+	        	<Search stocks={this.state.stockDataAPI}/>
 	        </div>
 		)
 	}
